LayoutTests/editing/selection/drag-list-item.html

<!DOCTYPE html>

<body>
<script src="../../resources/dump-as-markup.js"></script>
<div contenteditable="true">
<ol id="test">
<li id="one">one</li><li id="two">two</li><li id="three">three</li><li id="four">four</li>
</ol>
</div>

<script>
Markup.description('This test drags a selection of one <li> and drops them after another <li> in the first list. The same test is repeated in the second list but this time with two selected items instead of one. It verifies that the behavior is independent of how many fully selected items we move, i.e. fully selected <li> are moved as <li> and not just as plain text');

function editingTest(testElement, startSelectionElement, endSelectionElement, endSelectionElementOffset, targetElement) {

    var testElement = document.getElementById(testElement);
    var selection = window.getSelection();
    selection.setPosition(testElement, 0);

    var start = document.getElementById(startSelectionElement);
    var startx = start.offsetLeft;
    var starty = start.offsetTop + start.offsetHeight / 2;
    selection.setBaseAndExtent(start, 0, document.getElementById(endSelectionElement), endSelectionElementOffset);

    eventSender.mouseMoveTo(startx, starty);
    eventSender.mouseDown();

    eventSender.leapForward(200);

    var target = document.getElementById(targetElement);
    var targetx = target.parentNode.offsetLeft + target.offsetLeft + target.offsetWidth / 2;
    var targety = target.offsetTop + target.offsetHeight / 2;

    eventSender.mouseMoveTo(targetx, targety);
    eventSender.mouseUp();

    selection.collapse();
}

if (window.testRunner) {

    Markup.dump("test", "The original list looks like this. 'two' is going to be selected and pasted after 'four'");
    editingTest("test", "two", "two", 4, "four");
    Markup.dump("test", "'two' should appear as a new <li> after 'four'");

    document.getElementById("test").innerHTML = "<li id=\"one\">one</li><li id=\"two\">two</li>"
        + "<li id=\"three\">three</li><li id=\"four\">four</li>";
    Markup.dump("test", "The original list looks like this. 'two' and 'three' are going to be selected and pasted after 'four'");
    editingTest("test", "two", "three", 5, "four");
    Markup.dump("test", "'two' and 'three' should appear as <li> after 'four'");
}
</script>
</body>
</html>
