LayoutTests/editing/selection/drag-list-item-expected.txt

This test drags a selection of one <li> and drops them after another <li> in the first list. The same test is repeated in the second list but this time with two selected items instead of one. It verifies that the behavior is independent of how many fully selected items we move, i.e. fully selected <li> are moved as <li> and not just as plain text

The original list looks like this. 'two' is going to be selected and pasted after 'four':
| "
"
| <li>
|   id="one"
|   "one"
| <li>
|   id="two"
|   "two"
| <li>
|   id="three"
|   "three"
| <li>
|   id="four"
|   "four"
| "
"

'two' should appear as a new <li> after 'four':
| "
"
| <li>
|   id="one"
|   "one"
| <li>
|   id="two"
|   <br>
| <li>
|   id="three"
|   "three"
| <li>
|   id="four"
|   "four"
| <li>
|   id="two"
|   "two"
| "
"

The original list looks like this. 'two' and 'three' are going to be selected and pasted after 'four':
| <li>
|   id="one"
|   "one"
| <li>
|   id="two"
|   "two"
| <li>
|   id="three"
|   "three"
| <li>
|   id="four"
|   "four"

'two' and 'three' should appear as <li> after 'four':
| <li>
|   id="one"
|   "one"
| <li>
|   id="two"
|   <br>
| <li>
|   id="four"
|   "four"
| <li>
|   id="two"
|   "two"
| <li>
|   id="three"
|   "three"
